Add dedicated loader and plotter modules with tests
