+++ m3c2/visualization/__init__.py
@@ -1,1 +1,6 @@
- # Reporting/Plots
+"""Visualization utilities for the M3C2 project."""
+
+from .plot_comparedistances_service import PlotServiceCompareDistances
+
+__all__ = ["PlotServiceCompareDistances"]
+
