+++ tests/conftest.py
@@ -1,0 +1,9 @@
+"""Test configuration ensuring project package importability."""
+
+import os
+import sys
+
+ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+if ROOT not in sys.path:
+    sys.path.insert(0, ROOT)
+
